+++ mobile/android/android-components/components/service/firefox-accounts/src/test/java/mozilla/components/service/fxa/UtilsKtTest.kt
@@ -188,11 +188,4 @@ class UtilsKtTest {
             throw FxaPanicException("dunno")
         }
     }
-
-    @Test
-    fun `sync engine native name conversions`() {
-        assertEquals("passwords", SyncEngine.Passwords.toNativeString())
-        assertEquals("history", SyncEngine.History.toNativeString())
-        assertEquals("bookmarks", SyncEngine.Bookmarks.toNativeString())
-    }
 }