+++ mobile/android/android-components/samples/sync-logins/src/main/java/org/mozilla/samples/sync/logins/MainActivity.kt
@@ -17,6 +17,7 @@ import kotlinx.coroutines.Job
 import kotlinx.coroutines.async
 import kotlinx.coroutines.launch
 import mozilla.components.concept.sync.AccountObserver
+import mozilla.components.concept.sync.AuthType
 import mozilla.components.concept.sync.DeviceType
 import mozilla.components.concept.sync.OAuthAccount
 import mozilla.components.concept.sync.Profile
@@ -24,11 +25,13 @@ import mozilla.components.service.fxa.FirefoxAccount
 import mozilla.components.lib.fetch.httpurlconnection.HttpURLConnectionClient
 import mozilla.components.service.fxa.manager.FxaAccountManager
 import mozilla.components.service.fxa.DeviceConfig
+import mozilla.components.service.fxa.FxaAuthData
 import mozilla.components.service.fxa.ServerConfig
 import mozilla.components.service.fxa.SyncConfig
 import mozilla.components.service.fxa.SyncEngine
 import mozilla.components.service.fxa.sync.GlobalSyncableStoreProvider
 import mozilla.components.service.fxa.sync.SyncStatusObserver
+import mozilla.components.service.fxa.toAuthType
 import mozilla.components.service.sync.logins.AsyncLoginsStorageAdapter
 import mozilla.components.service.sync.logins.SyncableLoginsStore
 import mozilla.components.support.rusthttp.RustHttpConfig
@@ -42,18 +45,7 @@ const val CLIENT_ID = "3c49430b43dfba77"
 const val REDIRECT_URL = "https://accounts.firefox.com/oauth/success/$CLIENT_ID"
 
 class MainActivity : AppCompatActivity(), LoginFragment.OnLoginCompleteListener, CoroutineScope, SyncStatusObserver {
-    private val loginsStorage by lazy {
-        SyncableLoginsStore(
-                AsyncLoginsStorageAdapter.forDatabase(File(this.filesDir, "logins.sqlite").canonicalPath)
-        ) {
-            CompletableDeferred("my-not-so-secret-password")
-        }
-    }
-
-    init {
-        GlobalSyncableStoreProvider.configureStore("logins" to loginsStorage)
-    }
-
+    private lateinit var loginsStorage: SyncableLoginsStore
     private lateinit var listView: ListView
     private lateinit var adapter: ArrayAdapter<String>
     private lateinit var activityContext: MainActivity
@@ -63,7 +55,7 @@ class MainActivity : AppCompatActivity(), LoginFragment.OnLoginCompleteListener,
                 applicationContext,
                 ServerConfig.release(CLIENT_ID, REDIRECT_URL),
                 DeviceConfig("A-C Logins Sync Sample", DeviceType.MOBILE, setOf()),
-                SyncConfig(setOf(SyncEngine.Passwords))
+                SyncConfig(setOf(SyncEngine.PASSWORDS))
         )
     }
 
@@ -73,6 +65,7 @@ class MainActivity : AppCompatActivity(), LoginFragment.OnLoginCompleteListener,
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
         RustLog.enable()
         RustHttpConfig.setClient(lazy { HttpURLConnectionClient() })
 
@@ -88,6 +81,13 @@ class MainActivity : AppCompatActivity(), LoginFragment.OnLoginCompleteListener,
         adapter = ArrayAdapter(this, android.R.layout.simple_list_item_1)
         listView.adapter = adapter
         activityContext = this
+
+        loginsStorage = SyncableLoginsStore(
+            AsyncLoginsStorageAdapter.forDatabase(File(activityContext.filesDir, "logins.sqlite").canonicalPath)
+        ) {
+            CompletableDeferred("my-not-so-secret-password")
+        }
+        GlobalSyncableStoreProvider.configureStore(SyncEngine.PASSWORDS to loginsStorage)
 
         accountManager.register(accountObserver, owner = this, autoPause = true)
 
@@ -110,7 +110,7 @@ class MainActivity : AppCompatActivity(), LoginFragment.OnLoginCompleteListener,
         @Suppress("EmptyFunctionBlock")
         override fun onLoggedOut() {}
 
-        override fun onAuthenticated(account: OAuthAccount, newAccount: Boolean) {
+        override fun onAuthenticated(account: OAuthAccount, authType: AuthType) {
             accountManager.syncNowAsync()
         }
 
@@ -138,9 +138,11 @@ class MainActivity : AppCompatActivity(), LoginFragment.OnLoginCompleteListener,
         }
     }
 
-    override fun onLoginComplete(code: String, state: String, fragment: LoginFragment) {
+    override fun onLoginComplete(code: String, state: String, action: String, fragment: LoginFragment) {
         launch {
-            accountManager.finishAuthenticationAsync(code, state).await()
+            accountManager.finishAuthenticationAsync(
+                FxaAuthData(action.toAuthType(), code = code, state = state)
+            ).await()
             supportFragmentManager?.popBackStack()
         }
     }