+++ mobile/android/android-components/docs/changelog.md
@@ -49,6 +49,8 @@ permalink: /changelog/
   * `SyncConfig`'s `syncableStores` has been renamed to `supportedEngines`, expressed via new enum type `SyncEngine`.
   * `begin*` OAuthAccount methods now return an `AuthFlowUrl`, which encapsulates an OAuth state identifier.
   * `AccountObserver:onAuthenticated` method now has `authType` parameter (instead of `newAccount`), which describes in detail what caused an authentication.
+  * `GlobalSyncableStoreProvider.configureStore` now takes a pair of `Pair<SyncEngine, SyncableStore>`, instead of allowing arbitrary string names for engines.
+  * `GlobalSyncableStoreProvider.getStore` is no longer part of the public API.  
 
 # 11.0.0